hide output while tuning and stop device print for auto_koopman (#42)
